fix for vs warning

+++ inc/simple/btea.c
@@ -13,7 +13,7 @@ void btea(uint32_t *v, int n, uint32_t const key[4]) {
         do {
             sum += DELTA;
             e = (sum >> 2) & 3;
-            for (p=0; p<n-1; p++) {
+            for (p=0; p<(uint32_t)n-1; p++) {
                 y = v[p+1];
                 z = v[p] += MX;
             }
